requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
Pillow==10.3.0
python-decouple==3.8
